Minor refactor/cleanup in deployment processor
