+++ pkg/radrp/deployment/composite_error.go
@@ -1,0 +1,25 @@
+// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// ------------------------------------------------------------
+
+package deployment
+
+import "strings"
+
+// CompositeError represents an error containing multiple failures.
+type CompositeError struct {
+	Errors []error
+}
+
+func (ce *CompositeError) Error() string {
+	if len(ce.Errors) == 1 {
+		return ce.Errors[0].Error()
+	}
+
+	ss := make([]string, len(ce.Errors))
+	for i, e := range ce.Errors {
+		ss[i] = e.Error()
+	}
+	return "multiple errors: " + strings.Join(ss, ",")
+}
