+++ pkg/radrp/deployment/operations.go
@@ -1,0 +1,23 @@
+// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// ------------------------------------------------------------
+
+package deployment
+
+// DeploymentOperation represents an operation performed on a workload.
+type DeploymentOperation string
+
+const (
+	// None represents a workload that's unchanged in a deployment.
+	None DeploymentOperation = "none"
+
+	// DeleteWorkload represents deleting a workload from deployment.
+	DeleteWorkload DeploymentOperation = "delete"
+
+	// CreateWorkload represents creating a workload in deployment.
+	CreateWorkload DeploymentOperation = "create"
+
+	// UpdateWorkload represents updating a workload in deployment.
+	UpdateWorkload DeploymentOperation = "update"
+)
